tests/submission/MOSQUITO/required_attribute.html

<!DOCTYPE html>
<html>

  <head>
    <title>Required Attribute</title>
    <meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	<meta name="author" lang="fr" content="Fabrice Clari, Dimitri Bocquet">
    <script src="http://w3c-test.org/resources/testharness.js"></script>
    <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
  </head>

  <body>

  
  		<h1>Required Attribute</h1>
  		<div style="display: none">
  		<input type="text" required="required" />
		</div>
		
    <div id="log">
    </div>
	
	<script type="text/javascript">


		test(function() {assert_equals(document.getElementsByTagName("input")[0].getAttribute("required"), "required")}, "required attribute support on input element");

	</script>
	
  </body>
  
</html>
